refactor: rename stdlib to lang
